wip: config loads
