+++ __tests__/classes/config.ts
@@ -1,0 +1,7 @@
+import { Config } from "../../src/classes/config";
+
+describe("config class", () => {
+  it("loads config", () => {
+    expect(Config.general).toBeTruthy();
+  });
+});
