+++ src/config/i18n.ts
@@ -1,5 +1,5 @@
 export const DEFAULT = {
-  i18n: api => {
+  i18n: config => {
     return {
       // visit https://github.com/mashpie/i18n-node to see all configuration options
       // locale path can be configired from within ./config/api.js
