+++ src/config/plugins.ts
@@ -1,5 +1,5 @@
 export const DEFAULT = {
-  plugins: api => {
+  plugins: config => {
     /*
     If you want to use plugins in your application, include them here:
