+++ src/index.ts
@@ -1,5 +1,6 @@
-import { Api } from "./classes/api";
+// import { Api } from "./classes/api";
 export { Process } from "./classes/process";
+export { Config } from "./classes/config";
 export { Initializer } from "./classes/initializer";
 export { Connection } from "./classes/connection";
 export { Action } from "./classes/action";
@@ -8,4 +9,4 @@ export { Server } from "./classes/server";
 export { CLI } from "./classes/cli";
 export { ActionProcessor } from "./classes/actionProcessor";
 
-export const api = new Api();
+// export const api = new Api();
