+++ src/utils/sortGlobalMiddleware.ts
@@ -1,0 +1,15 @@
+/**
+ * Sorts an Array of Objects with a priority key
+ */
+export function sortGlobalMiddleware(
+  globalMiddlewareList: Array<any>,
+  middleware: Array<any>
+) {
+  globalMiddlewareList.sort((a, b) => {
+    if (middleware[a].priority > middleware[b].priority) {
+      return 1;
+    } else {
+      return -1;
+    }
+  });
+}
